Move test setup to top folder
