+++ _integration_tests/test/parallel_gateway_test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const should = require('should');
-const fixture = require('../application/dist/commonjs/process_engine_service_test_fixture')
+const fixture = require('../dist/commonjs/process_engine_service_test_fixture');
 
 const testTimeoutInMS = 5000;
